refactor: rearrange devserver files
